motogp_AI_strategist/requirements.txt
streamlit==1.35.0
pandas==1.5.3
numpy==1.23.5
matplotlib==3.10.3
pandasai==1.5.14
python-dotenv==1.1.1
openai==1.13.3